py/pyproject.toml
[project]
name = "llmproxy"
version = "0.1.4"
description = "Python client library for the LLMProxy server"
readme = "README.md"
authors = [
    { name = "Abdullah Bin Faisal" }
]
requires-python = ">=3.8"
dependencies = [
    "requests",
    "python-dotenv",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["llmproxy*"]